[SVG] Add CSS width and height support for <use> element
This CL introduces support for applying width and height as CSS
properties to SVG <use> element. The primary change is that the
viewport size for a <use> element referencing a <symbol> is
now determined by the computed width and height of the <use> element,
rather than the attributes or styles of the <symbol> or another <svg>
element.

Key Behavior Changes:

1. Enables width and height as presentation attributes on <use>.
2. Viewport resolution prioritizes <use>'s width and height over
<symbol>'s or another <svg> element.

Bug: 41413321
Change-Id: I638361ac3f2d2d30b1a282f47358aa2f19054644
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6646865
Reviewed-by: Divyansh Mangal <dmangal@microsoft.com>
Commit-Queue: Virali Purbey <viralipurbey@microsoft.com>
Reviewed-by: Dileep Maurya <dileepmaurya@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1484841}
NOKEYCHECK=True
GitOrigin-RevId: 37580d79be90601cdf03cb3f5e2b611f06120e58