[WPT Import] Add mechanism to mark CLs as exported
The WPT Importer process checks for exportable CLs.
Unlike the WPT Exporter, it cannot mark a CL as
exported after it has already been merged.

The importer is currently blocked because it has
identified two CLs for export, but one of them
cannot be applied cleanly.

Based on a suggestion from foolip@, this change
introduces a manual ignore list of Change-IDs to
resolve this blockage. This list allows the
importer to treat specific CLs as already exported.

This PR implements the ignore list and adds the
two problematic Change-IDs.

Other changes:
- Requires that each Change-ID on the list has a
  corresponding bug link for traceability.
- Added a mock for this new class. Similar to
  other mocks in the same directory

Bug: 430622361
Change-Id: Ifc081b04b23f6328834edaa9a86b81bcd2b361d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6720882
Commit-Queue: James Scott <jamescscott@google.com>
Reviewed-by: Johnny Stenback <jstenback@google.com>
Reviewed-by: Philip Jägenstedt <foolip@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1484901}
NOKEYCHECK=True
GitOrigin-RevId: 6b4304b612b0cf4c20d55eff205d5b16b1572ba6