[WebInstallAPI] Exclude feature from experimental web platform flag and exclude WebInstallAPI feature flag on Android
Currently, the WebInstallAPI feature is included in
chrome://flags/#enable-experimental-web-platform-features. The
WebInstallAPI feature has it's own feature flag so there is no need for
it to be included in the experimental web platform flag. This also fixes
the issue of the page crashing when
chrome://flags/#enable-experimental-web-platform-features flag is
enabled but chrome://flags/#web-app-installation-api flag is disabled
by default. This CL also excludes the WebInstallAPI feature flag on
Android and removes the last reference comment of crbug.com/381282538
that is no longer needed.

crbug.com/424497410 tracks supporting the API on Android in the future.

Bug: 429974686, 333795265, 381282538
Change-Id: Idc5b1b5b0d209ecfa4ef50e4e6cf8120546cac19
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6716192
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Kristin Lee <kristinlee@microsoft.com>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Lu Huang <luhua@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1485042}
NOKEYCHECK=True
GitOrigin-RevId: eef9e28d80153545142c79f019a4e0892d6121fd