Do not pause the HTML parser once it's resumed via user timing
This is part of the `HTMLParserYieldByUserTiming` experiment.

With `HTMLParserYieldByUserTiming`, normally the HTML parser is yielded
and then resumed by a user timing mark, but at the same time it's also
possible that the parser is resumed (this is no-op) and then paused due
to the wrong order of user timing events dispatched from the site. It
just makes a regression.

This CL prevents the parser from pausing in such case. It also adds
the UMA to measure how often it happens.

Bug: 425962649
Change-Id: I9aab6bf478fa4d7d1e669f412d0c37a5d8c715a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6721071
Reviewed-by: Jiacheng Guo <gjc@google.com>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Reviewed-by: Minoru Chikamune <chikamune@chromium.org>
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1485385}
NOKEYCHECK=True
GitOrigin-RevId: 6e5b7c9ed8d96e3796b28ed2356e936b66432bcc