Fix handling of <feConvolveMatrix divisor=0>
Per [1]:

  "If the specified divisor is 0 then the default value will be used
   instead.

   The initial value is the sum of all values in kernelMatrix, with the
   exception that if the sum is zero, then the divisor is set to 1."

If the `divisor` attribute was explicitly set to zero, we would not
apply the initial value, but rather fail the filter primitive.

Move the resolution of the initial value to FEConvolveMatrix and pass
zero if the attribute is not specified.

Remove the svg/filters/feConvolveMatrix-zero-divisor.svg test because
it appears redundant.

Guarded by the "SvgFeConvolveMatrixZeroDivisorBehavior" flag.

[1] https://drafts.fxtf.org/filter-effects-1/#element-attrdef-feconvolvematrix-divisor

Fixed: 431039946
Change-Id: I68abf37e3ee73f3f3cb3ea9b1cdb17d86d9e9fbf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6727535
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Cr-Commit-Position: refs/heads/main@{#1485532}
NOKEYCHECK=True
GitOrigin-RevId: e75cdaf538d09eaba04070051ec63db0a198d900