Set history index/length on view creation.
WebContentsImpl::SetHistoryIndexAndLengthForView is called when the
render view is initiated as part of a navigation, and results in an IPC
broadcast to update the associated render views' history index and
length. Not only is broadcasting the history/index length to unrelated
renderers redundant for creating a new render view, this also results in
a small window between WebViewImpl creation and the renderer receiving
the broadcast where the history index and length are set to invalid
values on the renderer side.

To fix that as well as reduce the number of unnecessary IPC calls going
out, this CL removes the broadcast call when a view is created and
instead sends the history index and length as part of the
CreateViewParams.

Note that WebViewImpl::SetHistoryListFromNavigation will be invoked for
all navigations when preparing to commit, including failed navigations,
so the values set during view creation will be overwritten (by current
values) at navigation commit time.

For keeping other renderer processes in sync, Navigator::DidNavigate
ensures that a broadcast will take place if the history length or index
changed as part of a separate renderer (associated with the same
RenderViewHost) committed a navigation.
https://source.chromium.org/chromium/chromium/src/+/main:content/browser/renderer_host/navigator.cc;drc=aee48482982cad62110c79eb1a9b1ffe831cd5b9;l=703-717

Example traces with navigation API:

Before:
https://ui.perfetto.dev/#!/?s=e88368b66eccef83b60cba0ab1f80b4e7f58ebf4

After:
https://ui.perfetto.dev/#!/?s=ded54002d4953fe092d06de283526976736c1ff7

Change-Id: Ic502b0c15bbe50481cde0a13669398b13f698376
Bug: 427982961
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6651954
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Liam Brady <lbrady@google.com>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1485578}
NOKEYCHECK=True
GitOrigin-RevId: f604c61efbc049404e3801b47f9899e5285e1679