[Trusted Types] Implement TT IFrame.srcdoc without StringContext.
This brings us in compliance with: https://html.spec.whatwg.org/#the-iframe-element

The [StringContext=] extendend IDL attribute is a Chrome-specific
mechanism to implement Trusted Types. TT is currently being integrated
into the HTML spec, and the spec editors decided against the
StringContext IDL attribute, and instead specify Trusted Types checks
"manually" in the spec. In some instances -- but not here -- the
difference between the IDL-based and the "manual" TT checks is observable, due to the when the TT check is being made.

This CL re-implements the Trusted Types check of HTMLIFrameElement.srcdoc.

Spec: https://html.spec.whatwg.org/#the-iframe-element
Bug: 330516530
Change-Id: I9617ae6778585482a46878dbe52ec8856f1301fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6632847
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Commit-Queue: Daniel Vogelheim <vogelheim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1485770}
NOKEYCHECK=True
GitOrigin-RevId: 84fc9a73306b91a597060583b28357790c0b0a0d