[Trusted Types] Remove StringContext extended IDL attribute.
This removes the [StringContext=] extended IDL attribute. It
was introduced to implement large parts of Trusted Types.
Recently, the TT spec was being integrated into HTML proper, and
as part of that work a decision was made to implement Trusted
Types using union types, instead of an IDL annotation. This
CL completes a refactoring of Chrome's TT implementation, to
bring it in line with the updated HTML spec.

This is a deletion-only CL.

Spec: https://github.com/whatwg/html/pull/10286
Bug: 330516530
Change-Id: Id7444510658d6fa3231fa34ca9cffa5e532e8efc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6633622
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Commit-Queue: Daniel Vogelheim <vogelheim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1486260}
NOKEYCHECK=True
GitOrigin-RevId: c97415436cf10cd52a79e1aa0a76755d5b98f941