[gap-decorations] Interpolation with repeaters for *-rule-width
This CL implements Interpolation for *-rule-width for repeaters.

In order to handle repeaters, we add a new InterpolableType
`InterpolableGapDataRepeater` which handles interpolation for repeaters
by essentially keeping a member InterpolableList of the values in the
repeater (for now just `InterpolableLength`) and interpolating those
individually through that type.

The overall structure of interpolation for gap decorations will be as
follows:

`CSSGapLengthListInterpolationType` handles interpolation for
`GapDataList` We then hand off the actual interpolation with an
`InterpolableList` of `InterpolableLength`,
`InterpolableGapDataRepeater` (which itself contains an
`InterpolableList` of `InterpolableLengths`) and `InterpolableColor`.

InterpolableGapDataRepeater is templated, in order to handle for
interpolation for column/row-rule-color. This will be implemented in a
follow-up CL.

The changes to `rule-width-interpolation-multiple-values-*` are because
for now we are implementing mirroring the behavior of `grid-template`
interpolation (which can be a list of values with repeaters), in which
we only interpolate if the length of the lists match. If this changes
after spec discussions then the test will be updated.

https://www.w3.org/TR/css-gaps-1/#lists-repeat

Bug: 357648037, 419066541

Change-Id: Ia43396607d0bbe3a6266571df31ae36feb882875
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6685488
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Javier Contreras <javiercon@microsoft.com>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1486359}
NOKEYCHECK=True
GitOrigin-RevId: fa3e4567ef10e752185f7bbe561bb6acf26a9ad8