Include transforms when subtracting layerX/layerY container position.
It doesn't make sense not to include transforms, because before
this offset subtraction the value already includes transforms.

See:

https://github.com/w3c/uievents/issues/398#issuecomment-3070325191

This CL may impact web compat, but I don't think it will be
significant because layerX and layerY are not very interoperable
or widely used, because they are not very useful. This CL is just
a step towards interop in this particular corner case.

Change-Id: I1b2f0dbf34b44dce7bde17e17cb315b79b6cada8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6733649
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Chris Harrelson <chrishtr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1486571}
NOKEYCHECK=True
GitOrigin-RevId: 5a32b6a6abebde6c5762653eff3587619682f524