Make fetch retry behavior for different network errors not web-visible
This reverts commit ddf7f5c6e19d7b1258b0b325baec800ea57c20ed
(crrev.com/c/6703481) and makes it so that network errors
encountered by a fetch with retry options set will only be
surfaced at kMaxAge.

Reason: The standardization discussion at https://github.com/whatwg/fetch/issues/1838#issuecomment-3069011754
concluded that it's actually ok to retry on certain errors, but
not make the difference visible to script, by delaying the resolving
of all errors until we reach kMaxAge.

This CL makes it so that when getting an error for a fetch with
retry options set, and a retry will no longer be attempted, we delay
the processing of the error until the max age set in the retry
options if the renderer is connected. If the renderer is
disconnected, no information can be leaked, so we just continue
processing the error normally.

This CL also makes the maxAge option to be required.

Revert is in PS1: https://chromium-review.googlesource.com/c/chromium/src/+/6734418/1

Actual addition by this CL is the diff from PS1..8:
https://chromium-review.googlesource.com/c/chromium/src/+/6734418/1..8

Bug: 417930271
Original change's description:
> Make fetch retry attempt retry on all network errors
>
> Per feedback in https://github.com/whatwg/fetch/issues/1838, only
> attempting to retry on transient network errors and not retrying on e.g.
> policy errors might introduce a privacy problem if the requester can
> infer whether a retry happened from the time it took for the fetch to
> resolve.
>
> This CL makes it so that all network errors are eligible for retry, so
> that there's no way to tell the difference between errors using fetch
> retry.
>
> Bug: 417930271
> Change-Id: I601fac0fba34decaae66e20b73be9d5ed8fcf29d
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6703481
> Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
> Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1482443}

Bug: 417930271
Change-Id: Ie01498b51a7d94e6592ce4e32ffe5e7dcc8776d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6734418
Reviewed-by: Ming-Ying Chung <mych@chromium.org>
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1487448}
NOKEYCHECK=True
GitOrigin-RevId: a06e26afba9745b1d14919279b49ab816f9d76ea