[refactor] rm should_skip_preload_scan_
The logic is simplified to directly check the ShouldSkipPreloadScan
condition for creating the `HTMLResourcePreloader`.

Subsequent operations that depend on this check are now simplified to verify the existence of the preloader instance. This ensures that the decision to create a preloader and the decision to use it are always consistent, removing redundant state and improving clarity.

This should be a safe refactor because:
1. it is safe not to create preloader_ in the case that `should_skip_preload_scan_` is true.
preloader_ is used to fetch the preload resources generated by
`preload_scanner_`, `insertion_preload_scanner_` and
`background_scanner_`, and these scanners will not be created if
`should_skip_preload_scan_` is set to false. So preloader_ is actually
not used when should_skip_preload_scan_ is false, and we do
not need to create it.

2. (preloader_ && !should_skip_preload_scan_)'s shortcut is
if(should_skip_preload_scan_) preloader_ = nullptr;

Bug: N/A
Change-Id: I9269829e9db4e54606dbdfd1610361485d28b62c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6683498
Commit-Queue: Lingqi Chi <lingqi@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1488027}
NOKEYCHECK=True
GitOrigin-RevId: 40a76f2e822e7ed6fb3c42014e9c24820d6c4f3d