[Gap Decorations]: Suppress gaps for column flex containers
This CL introduces the suppression of row gaps in column flex containers
when they are split across multiple fragments. It begins by augmenting
the existing `UpdateOffsetAdjustmentForSuppressedRowGap` function to
support both row and column flex containers.

The key difference in logic lies when a break occurs that is not forced.
In row flex containers, we can assert that the row gap
(`gap_between_lines_`) will be greater than or equal to the available
space. However, this assumption does not hold true for column flex
containers.

For example, when a monolithic item cannot fit in the remaining space,
it is deferred to the next fragment. In such cases, the row gap
(`gap_between_items_`) may already have been accounted for in the
current fragmentainer, simply because it fit within the available space.
This behavior requires us to scope suppression only when the row gap is
greater than the available space.

Bug: 357648037, 394042462
Change-Id: I6af564ec778961db07a8d3f74e6ffedc9c94b6ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6666279
Reviewed-by: Alison Maher <almaher@microsoft.com>
Commit-Queue: Sam Davis Omekara <samomekarajr@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1488313}
NOKEYCHECK=True
GitOrigin-RevId: 0148cb3321ebb93320096cf986525981895b5306