[a11y] Ensure empty list items link to list markers
This commit introduces a fix for, as well as a new
accessibility test
(AccessibilityNextOnLineEmptyListItem) to verify
that inline text boxes, even when empty and
containing only a new line, correctly point back
to their associated list marker node.

This fix ensures that screen readers can properly
parse the list marker and the subsequent new line
as being on the same line. This results in the
correct utterance of "bullet" for empty list
items, preventing the incorrect announcement of
"new line" which would occur if the connection to
the bullet marker were lost.

Change-Id: I8961bfb192d50ebe65dfb4c28e8dcf099fca1c66
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6762918
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Kevin Ellis <kevers@chromium.org>
Commit-Queue: Lucas Radaelli <lucasradaelli@google.com>
Cr-Commit-Position: refs/heads/main@{#1488360}
NOKEYCHECK=True
GitOrigin-RevId: 20fb588d4bfc65198c59f3fcf12f2837793fa214