[DC] Refactor DiscoverDigitalIdentityCredentialFromExternalSource()
This is a no-op that inlines the lambda into the loop in which it is
invoked.

The lambda has been there due to some legacy more complex logic. The
logic has been simplified in https://crrev.com/c/6760153 but the lambda
has been left to simplify the review.

Bug: 431999487
Change-Id: Ie79f3c22370866b56a1da37f5377b90396f3f659
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6766955
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Commit-Queue: Mohamed Amir Yosef <mamir@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1488699}
NOKEYCHECK=True
GitOrigin-RevId: db162a61a0b3a0e92ab4f838d6e97d4ef63f97d5