Reland "Implement a simple mechanism for downstream forks to customize runtime-enabled-features.json5"
Ensure file_paths are only updated if there are items that are added/changed.

Original Description:

This is a change to the runtime-enabled-featured build infrastructure to
look for an embedder specific file and if present allows its contents to
override anything set in runtime_enabled_features.json5.

I've included a few examples of how this would work in
runtime_enabled_features_valid.override.json5

I've also confirmed that if a feature is configured the override file,
that the tests inherit this behavior.

e.g. I changed CSSScrollTargetGroup in the overridden file to set the
status to experimental. This resulted in expected test failures:
https://ci.chromium.org/ui/p/chromium/builders/try/linux-rel/b8709621089823682465/test-results?q=ExactID%3Aninja%3A%2F%2F%3Ablink_web_tests%2Fvirtual%2Fstable%2Fwebexposed%2Fcss-properties-as-js-properties.html+VHash%3A32f081785efe9631&clean=

Bug: 428882127
Change-Id: I5b5f79aae74c9f43f19fd79db0e3dc86774b83c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6731644
Reviewed-by: Rick Byers <rbyers@chromium.org>
Commit-Queue: Mike Jackson <mjackson@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1489092}
NOKEYCHECK=True
GitOrigin-RevId: 0088ecaf5559aade6a67b097fda04cd60d185d9a