[Masonry] Use the correct margin value when calculating track size
The `ComputeMarginsFor` method we use when calculating track sizes
already translates the margin values to be local to the masonry
container, so we can assume the block/inline size of the margin returned
logical to the container regardless of if an item is orthogonal or not.

This change adds testing for some new scenarios:
* margins are added to the inline-axis of an orthogonal item in a column layout, to ensure that track sizing is done correctly even with margins on an orthogonal item
* items are placed after an orthogonal
item to ensure that the running position of orthogonal items is
calculated correctly.
* margins are placed between orthogonal items to ensure that running position is calculated correctly.

Bug: 343257585
Change-Id: I35437e48d3c049581ac41a9a7a5e57b386956dd8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6761596
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Commit-Queue: Celeste Pan <celestepan@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1489103}
NOKEYCHECK=True
GitOrigin-RevId: 46d8f823cf75bc22c4e11244a38fee6a1b9498b2