[Blink] Move IsHibernating() to be method on CanvasRenderingContext
We will be moving CanvasHibernationHandler to be owned by and internal
to CanvasRenderingContext2D, since it is Canvas2D-specific. This CL
takes the first step in that direction by moving
CanvasRenderingContextHost::IsHibernating() to be a virtual method on
CanvasRenderingContext, with CanvasRenderingContext2D taking the
implementation that is currently in HTMLCanvasElement. We adjust all
callsites, which in CanvasRenderingContextHost/HTMLCanvasElement entails
adding conditionals that guard for the context being present. Adding
those conditionals is not a behavioral change as it is only
CanvasRenderingContext2D that initiates hibernation [1], and thus if
there is no context present then it is not possible for the canvas to
be hibernating.

[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/platform/graphics/canvas_hibernation_handler.h;l=76;bpv=1;bpt=1?q=InitiateHibernationIf&ss=chromium

Bug: 352263194
Change-Id: I42e4528fda5168fe6b47b28a09b742411b1cb734
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6769980
Reviewed-by: Saifuddin Hitawala <hitawala@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1489384}
NOKEYCHECK=True
GitOrigin-RevId: 2fa3fed4400a997408c5e0d2b4d461b3d6b12b73