[Blink] Move hibernation handler to be owned by CanvasRenderingContext2D
This CL moves the ownership of CanvasHibernationHandler from
HTMLCanvasElement to CanvasRenderingContext2D. Hibernation is
Canvas2D-specific, so it makes sense for knowledge of the handler to be
localized within the 2D context. The immediate motivation is that the
handler interacts with the CanvasResourceProvider for canvas2D, which
we're also in the process of moving to the 2D context.

In addition to moving ownership of the handler to the 2D context, this
CL moves implementation of the delegate interface from HTMLCanvasElement
to CanvasRenderingContext2D. It's easiest to make both of those changes
at the same time to avoid niggling lifetime issues that would otherwise
pop up due to the fact that the context and the canvas can be dropped
in either order. In this initial CL making the move, we essentially have
the context implement the delegate interface by just calling back to the
canvas to make it obvious that there are no behavioral changes. We'll
streamline this over time.

Bug: 352263194
Change-Id: Iaf1165e553d532f0428eebe11b0c164b99cd9a45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6770241
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Saifuddin Hitawala <hitawala@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1489396}
NOKEYCHECK=True
GitOrigin-RevId: 342ce56438a9ba40dcd31cadc664a23ef1c9812e