[EditContext] Return early in SetSelection for unchanged bounds
During `RenderWidgetHostViewCocoa::setTextSelectionText` text
replacements up to the previous character are either accepted or
rejected based on the current update text and range.

In `EditContext::SetSelection`, the call to
`LocalFrameClient::DidChangeSelection` triggers a selection change
notification to the browser, which in turn calls setTextSelectionText.
This sequence is triggered twice for the same selection—first during
`EditContext::InsertText`, and again during
`EditContext::UpdateSelection`. When a text replacement suggestion is
active, the duplicate DidChangeSelection call causes the browser to
incorrectly assume that the suggestion has been accepted. As a result,
any key press following an active suggestion leads to the replacement
being accepted, even if it shouldn't be.

This CL addresses the issue by adding an early return in
`EditContext::SetSelection` when the selection bounds haven't changed,
to avoid duplicate selection change notifications to the browser.

Bug: 428112177
Change-Id: I42000a81ba7305f02ea768964377064b13acd9fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6763992
Reviewed-by: Dan Clark <daniec@microsoft.com>
Commit-Queue: Ashish Kumar <ashishkum@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1489935}
NOKEYCHECK=True
GitOrigin-RevId: 25395e79d35aa7354e1e5644f92f4f93c48dca9f