[Blink] Eliminate calling GpuCanvasContext::IsAccelerated()
The one callsite of this method for WebGPU is here [1]. That method's
one production callsite is here [2]. The WebGPU implementation of
IsAccelerated() checks whether the context's CanvasResourceProvider
exists and is accelerated. This works as expected when the WebGPU canvas
is an HTMLCanvasElement, as in that case
PaintRenderingResultsToSnapshot() will have been called via
HTMLCanvasElement::GetSourceImageForCanvas() just prior to the call to
[2], and it will have created an accelerated CRP iff the
StaticBitmapImage that was returned by calling Snapshot() on that CRP
is accelerated. However, it does not work as intended when the WebGPU
canvas is an offscreen canvas: it will never accelerate, as
GpuCanvasContext never creates a CRP when used as an offscreen canvas
(whether a placeholder is present [3] or not).

This CL changes [2] to use the info of whether the created image is
texture-backed as the source of truth for whether a WebGPU canvas is
accelerated. This leaves the current behavior unchanged for the case
where the WebGPU canvas is an HTMLCanvasElement and aligns the behavior
for the case where the WebGPU canvas is an OffscreenCanvas to match the
current behavior for the case of it being an HTMLCanvasElement.

[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/html/canvas/canvas_rendering_context_host.cc;l=186;drc=d0cdb873a9f5d9b6ef2696f65cdd80b619240b14;bpv=1;bpt=1
[2] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/html/canvas/html_canvas_element.cc;l=1924;drc=d0cdb873a9f5d9b6ef2696f65cdd80b619240b14
[3] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/modules/webgpu/gpu_canvas_context.cc;l=328-339?q=GpuCanvasContext::Push&ss=chromium

Bug: 352263194
Change-Id: Ia413ed17e425a94bb9e046ecdc6242b719360df4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6772821
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1489955}
NOKEYCHECK=True
GitOrigin-RevId: a31388a0672c463693b2cc636ccecbeaef86de02