Carry the information about nested intermediate result in css math node
When we have a math node with intermediate result (e.g. 10px / 1px)
as a whole operation or just a part of it, we can't simplify that node,
as it would result in a crash, since we can only simplify number/typed
node relations (e.g. 10px * 2 -> 20px).

This CL bakes this information in CSSMathExpressionOperation node and
carries it around to check if simplification conditions.

Fixed: 432385233
Change-Id: I0674efe45f9ae0274d0f2880ff0017663d27b7c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6771850
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Daniil Sakhapov <sakhapov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1490015}
NOKEYCHECK=True
GitOrigin-RevId: c9a863cca434ec54b9157434856b3ca6456d266a