[owners] Remove weizhong@google.com from third_party/blink/web_tests/external/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://chromium-review.googlesource.com/q/weizhong@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Iab84c3c8158167fa45e8b99b16580f68b9be3bde
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6652554
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Philip Jägenstedt <foolip@chromium.org>
Commit-Queue: Philip Jägenstedt <foolip@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1490042}
NOKEYCHECK=True
GitOrigin-RevId: c12a3a1b51225d775dda6622490152a48cdf92ae