ResourceLoader's cancel timer should not be frozen
This timer only starts when no ResourceClient is observing the load
any longer. We've seen issues with cleanup not happening at all if
we navigate away with resource loads in progress but deferred, so
ensure that the timer for that cleanup is unfreezable.

Bug: 41161335
Change-Id: I60b6f9ef5eb2344e98f7554280afe4cac9e1a3de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6767556
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Nate Chapin <japhet@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1490435}
NOKEYCHECK=True
GitOrigin-RevId: 5cc7932eff66dd2ae4773f78638e69d778c0f582