[Blink] Check whether source image is texture-backed for Canvas2D
The method in question is here [1]. That method's
one production callsite is here [2]. The Canvas2D implementation of
IsAccelerated() checks whether the context's CanvasResourceProvider
exists and is accelerated. For Canvas2D in the flow in question this is
equivalent to checking whether the passed-in image is texture-backed,
as that image will have been generated via
HTMLCanvasElement::GetSourceImageForCanvas() just prior to the call to
[2], and both HTMLCanvasElement and OffscreenCanvas implement
GetSourceImageForCanvas() for canvas2D by calling GetImage() on the
context [3][4], which for (Offscreen)CanvasRenderingContext2D creates
the CRP and returns a snapshot from it.

This CL changes [2] to use the info of whether the created image is
texture-backed as the source of truth in this flow for whether a source
image from a canvas2D canvas is accelerated. This paves the way for
fixing the case where the source is an ImageBitmap canvas in followup
and eliminating this flow ever calling
CanvasRenderingContext::IsAccelerated().

[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/html/canvas/canvas_rendering_context_host.cc;l=186;drc=d0cdb873a9f5d9b6ef2696f65cdd80b619240b14;bpv=1;bpt=1
[2] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/html/canvas/html_canvas_element.cc;l=1924;drc=d0cdb873a9f5d9b6ef2696f65cdd80b619240b14
[3] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/html/canvas/html_canvas_element.cc;l=1991?q=HTMLCanvasElement::GetSourceIma&ss=chromium
[4] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/offscreencanvas/offscreen_canvas.cc;l=274?q=OffscreenCanvas::GetSou&ss=chromium

Bug: 352263194
Change-Id: Ie06ff2d7b0b1d8232bdadc1769cc48a4c5355623
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6776637
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1490629}
NOKEYCHECK=True
GitOrigin-RevId: d97c215891b0199ffbf165bf92fa99ae9d8aa475