[text-spacing-trim] Support fallback (partial)
Before this patch, `text-spacing-trim` may fail to apply when
character pairs are in different fonts. This may occurs when,
for example, web fonts provider chunks fonts to limit the size
of each font to download.

To support this, `HanKerning` enables a cache of `CharType` in
the shaping segment if fallback occurs after it has applied
font features.

Note that this is a partial fix. Because it is caching the
previously computed values, it fails if fonts to apply
features appear later in the cascading chain. This case will
be addressed in future patches.

All logic changes are under a runtime flag
`TextSpacingTrimFallback`.

Bug: 431660829
Change-Id: I82cbd1b378a35cd7471ec6a8bbb5545972bcc034
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6772946
Reviewed-by: Kent Tamura <tkent@chromium.org>
Auto-Submit: Koji Ishii <kojii@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1490699}
NOKEYCHECK=True
GitOrigin-RevId: ace0e61671c0c5babe23418d7f0dec37c4a1d025